refactor: images api refactoring
